// src/app/list-cars/interface/cars.interface.ts
export interface Vehiculo {
	id: number;
	administrador_dni: number;
	categoria: string;
	imagen: string;
	color: string;
	estado_id: number;
	kilometraje: number;
	marca: string;
	modelo: string;
	numero_asientos: number;
	placa: string;
	soat: string;
	transmision: string;
	price: string;
	ocupantes: string;
}